+++ Scraps.Next/Controls/MainControl.Designer.cs
@@ -32,14 +32,17 @@ namespace Scraps.Next.Controls
         {
             this._RunnerButton = new System.Windows.Forms.Button();
             this._MainViewLog = new System.Windows.Forms.RichTextBox();
+            this._AlertLabel = new System.Windows.Forms.Label();
             this.SuspendLayout();
             // 
             // _RunnerButton
             // 
+            this._RunnerButton.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
             this._RunnerButton.Image = global::Scraps.Next.Resources.Images.control;
-            this._RunnerButton.Location = new System.Drawing.Point(9, 9);
+            this._RunnerButton.Location = new System.Drawing.Point(6, 371);
+            this._RunnerButton.Margin = new System.Windows.Forms.Padding(0, 0, 6, 0);
             this._RunnerButton.Name = "_RunnerButton";
-            this._RunnerButton.Size = new System.Drawing.Size(83, 24);
+            this._RunnerButton.Size = new System.Drawing.Size(97, 24);
             this._RunnerButton.TabIndex = 0;
             this._RunnerButton.Text = "Start";
             this._RunnerButton.TextAlign = System.Drawing.ContentAlignment.MiddleRight;
@@ -52,25 +55,39 @@ namespace Scraps.Next.Controls
             | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
             this._MainViewLog.BackColor = System.Drawing.Color.Black;
+            this._MainViewLog.BorderStyle = System.Windows.Forms.BorderStyle.None;
             this._MainViewLog.Font = new System.Drawing.Font("Cascadia Code", 9.75F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point);
             this._MainViewLog.ForeColor = System.Drawing.Color.White;
-            this._MainViewLog.Location = new System.Drawing.Point(98, 9);
+            this._MainViewLog.Location = new System.Drawing.Point(6, 9);
             this._MainViewLog.Name = "_MainViewLog";
             this._MainViewLog.ReadOnly = true;
             this._MainViewLog.ScrollBars = System.Windows.Forms.RichTextBoxScrollBars.Vertical;
-            this._MainViewLog.Size = new System.Drawing.Size(720, 380);
+            this._MainViewLog.Size = new System.Drawing.Size(812, 359);
             this._MainViewLog.TabIndex = 2;
             this._MainViewLog.Text = "";
+            // 
+            // _AlertLabel
+            // 
+            this._AlertLabel.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left) 
+            | System.Windows.Forms.AnchorStyles.Right)));
+            this._AlertLabel.Location = new System.Drawing.Point(112, 371);
+            this._AlertLabel.Name = "_AlertLabel";
+            this._AlertLabel.Size = new System.Drawing.Size(706, 24);
+            this._AlertLabel.TabIndex = 3;
+            this._AlertLabel.Text = "label1";
+            this._AlertLabel.TextAlign = System.Drawing.ContentAlignment.MiddleLeft;
             // 
             // MainControl
             // 
             this.AutoScaleDimensions = new System.Drawing.SizeF(7F, 15F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
             this.BackColor = System.Drawing.Color.White;
+            this.Controls.Add(this._AlertLabel);
             this.Controls.Add(this._MainViewLog);
             this.Controls.Add(this._RunnerButton);
+            this.Margin = new System.Windows.Forms.Padding(0);
             this.Name = "MainControl";
-            this.Padding = new System.Windows.Forms.Padding(6);
+            this.Padding = new System.Windows.Forms.Padding(6, 6, 6, 3);
             this.Size = new System.Drawing.Size(827, 398);
             this.Load += new System.EventHandler(this.MainControl_Load);
             this.ResumeLayout(false);
@@ -81,5 +98,6 @@ namespace Scraps.Next.Controls
 
         private Button _RunnerButton;
         private RichTextBox _MainViewLog;
+        private Label _AlertLabel;
     }
 }